feat: add rect and restore opacity working for some reason

+++ src/fabric/CircleNativeComponent.js
@@ -22,7 +22,7 @@ type SvgNodeCommonProps = $ReadOnly<{|
   name?: string,
   // It has to be Double on Android no to clash with BaseViewManager method
   // TODO: make it work with Float
-  opacity?: WithDefault<Double, 1.0>,
+  opacity?: WithDefault<Float, 1.0>,
   matrix?: $ReadOnlyArray<Float>,//CGAffineTransform,
   // transform?: ____TransformStyle_Internal, // CATransform3D, custom handling
   mask?: string,
